fruit_api/app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
tensorflow==2.17.0
pillow==10.4.0
